requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1 